# Add support for QRWKV6 hybrid models & slight optimization for RWKV6