--- a/ggml/include/ggml.h
+++ b/ggml/include/ggml.h
@@ -513,6 +513,7 @@ extern "C" {
         GGML_OP_GET_REL_POS,
         GGML_OP_ADD_REL_POS,
         GGML_OP_RWKV_WKV6,
+        GGML_OP_GATED_LINEAR_ATTENTION,
 
         GGML_OP_UNARY,
 
@@ -1871,6 +1872,14 @@ extern "C" {
             struct ggml_tensor  * r,
             struct ggml_tensor  * tf,
             struct ggml_tensor  * td,
+            struct ggml_tensor  * state);
+
+    GGML_API struct ggml_tensor * ggml_gated_linear_attn(
+            struct ggml_context * ctx,
+            struct ggml_tensor  * q,
+            struct ggml_tensor  * k,
+            struct ggml_tensor  * v,
+            struct ggml_tensor  * g,
             struct ggml_tensor  * state);
 
     // custom operators