--- a/ggml/include/ggml.h
+++ b/ggml/include/ggml.h
@@ -513,7 +513,7 @@ extern "C" {
         GGML_OP_GET_REL_POS,
         GGML_OP_ADD_REL_POS,
         GGML_OP_RWKV_WKV6,
-        GGML_OP_GATED_LINEAR_ATTENTION,
+        GGML_OP_GATED_LINEAR_ATTN,
 
         GGML_OP_UNARY,
 
@@ -1876,11 +1876,12 @@ extern "C" {
 
     GGML_API struct ggml_tensor * ggml_gated_linear_attn(
             struct ggml_context * ctx,
-            struct ggml_tensor  * q,
             struct ggml_tensor  * k,
             struct ggml_tensor  * v,
+            struct ggml_tensor  * q,
             struct ggml_tensor  * g,
-            struct ggml_tensor  * state);
+            struct ggml_tensor  * state,
+            float scale);
 
     // custom operators