--- a/ggml/src/ggml-cuda/gla.cuh
+++ b/ggml/src/ggml-cuda/gla.cuh
@@ -1,0 +1,3 @@
+#include "common.cuh"
+
+void ggml_cuda_op_gated_linear_attn(ggml_backend_cuda_context & ctx, ggml_tensor * dst);