… README.md
… gguf-py/gguf/constants.py
… gguf-py/gguf/gguf_writer.py
… gguf-py/gguf/tensor_mapping.py
… src/llama-arch.cpp
… src/llama-arch.h
… src/llama-hparams.cpp
… src/llama-hparams.h
… src/llama-model.cpp
… src/llama-model.h
… src/llama-quant.cpp
… src/llama.cpp
… tests/test-backend-ops.cpp