@MollySophia 2025-01-10
Self-reviewed again and fixed a merge issue after the refacor that makes RWKV6 fail to run with fused lerp weights.
I think this PR is ready to go.